create card-viewer component, more tweaks
