+++ packages/cardhost/app/routes/cards/card/edit/layout.js
@@ -1,4 +1,2 @@
 import CardModelRoute from '../card-model-route';
-export default class CardEditLayoutRoute extends CardModelRoute {
-  templateName = 'cards/card/view';
-}
+export default class CardEditLayoutRoute extends CardModelRoute {}
