+++ packages/cardhost/app/templates/cards.hbs
@@ -6,35 +6,6 @@
     {{if this.cardstackSession.isAuthenticated "logged-in" "logged-out"}}
     {{this.themerClasses}}"
     data-test-cardhost-cards>
-    {{#unless this.onThemerRoute}}
-      <CardhostTopEdge @mode={{mode}}>
-        {{#if (or (eq mode "edit") (eq mode "layout"))}}
-          {{#let this.target.currentRoute.parent.attributes as |card|}}
-            <div class="view-selector">
-              <LinkTo
-                @route="cards.card.layout"
-                @class="view-selector--mode-link {{mode}}"
-                @model={{card.name}}
-                data-test-view-selector="layout"
-              >
-                {{svg-jar "layout" width="18px" height="17px"}}
-                Layout
-              </LinkTo>
-              <LinkTo
-                @route="cards.card.edit"
-                @class="view-selector--mode-link"
-                @model={{card.name}}
-                data-test-view-selector="field"
-              >
-                {{svg-jar "form" width="18px" height="17px"}}
-                Field
-              </LinkTo>
-            </div>
-          {{/let}}
-        {{/if}}
-      </CardhostTopEdge>
-    {{/unless}}
-
     {{outlet}}
   </div>
 {{/let}}
