+++ packages/cardhost/app/templates/cards/card/edit/layout/index.hbs
@@ -1,0 +1,1 @@
+<CardViewer @card={{this.model}} />
