+++ packages/core/addon/templates/components/card-renderer.hbs
@@ -29,7 +29,7 @@ Ember Concurrency into the rendering of the card component here.
               {{svg-jar "event" width="18px" height="20px"}}
             {{/if}}
             <span class="card-renderer-isolated--header-title hide-in-percy">{{or @card.name "Blank Card"}}</span>
-            {{#if (eq @mode "view")}}
+            {{#if (eq this.mode "view")}}
               <LinkTo
                 @route="cards.card.edit.fields"
                 @model={{@card}}
